Add possibility to *not* backup an existing trajectory file (avoid an old.traj.bak file).

+++ ase/io/trajectory.py
@@ -18,7 +18,7 @@ class PickleTrajectory:
     write_momenta = True
     
     def __init__(self, filename, mode='r', atoms=None, master=None,
-                 write_first_image=True):
+                 write_first_image=True, backup=True):
         """A PickleTrajectory can be created in read, write or append mode.
 
         Parameters:
@@ -50,19 +50,21 @@ class PickleTrajectory:
             argument is given, processes where it is True will write.
 
         write_first_image=True:
-        
             If this argument is True and atoms is specified, the atoms
             are written to the file immediately.  This is intended to
             write the initial configuration in a simulation.  Note
             that in append mode you probably want to set this to
             False, as the first configuration is likely to already be
             in the file.
-        
+
+        backup=True:
+            Use backup=False to disable renaming of an existing file.
         """
         self.offsets = []
         if master is None:
             master = (rank == 0)
         self.master = master
+        self.backup = backup
         self.set_atoms(atoms)
         self.open(filename, mode)
 
@@ -89,7 +91,7 @@ class PickleTrajectory:
         elif mode == 'w':
             if self.master:
                 if isinstance(filename, str):
-                    if os.path.isfile(filename):
+                    if self.backup and os.path.isfile(filename):
                         os.rename(filename, filename + '.bak')
                     self.fd = open(filename, 'wb')
             else:
